refactor: separate UserFixture from ThemeBuildFixture Related to: #137
